+++ config/jest/raf.js
@@ -1,5 +1,0 @@
-const raf = global.requestAnimationFrame = function(callback) {
-  setTimeout(callback, 0);
-};
-
-export default raf;