+++ config/jest/setupTests.js
@@ -1,6 +1,0 @@
-import './raf'; // leave this before the enzyme imports
-
-import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-Enzyme.configure({ adapter: new Adapter() });