+++ config/jest/setupTestsAfterEnv.js
@@ -1,6 +1,0 @@
-window.localStorage = {
-  getItem: jest.fn(() => '{}'),
-  setItem: jest.fn(),
-  removeItem: jest.fn(),
-  clear: jest.fn()
-};