+++ jest.config.js
@@ -1,13 +1,11 @@
 module.exports = {
   collectCoverageFrom: ['src/**/*.{js,jsx}'],
-  snapshotSerializers: ['enzyme-to-json/serializer'],
-  setupFiles: ['<rootDir>/config/jest/setupTests.js'],
   setupFilesAfterEnv: ['<rootDir>/config/jest/setupTestsAfterEnv.js'],
   testMatch: [
     '<rootDir>/src/**/__tests__/**/*.js?(x)',
     '<rootDir>/src/**/?(*.)(spec|test).js?(x)'
   ],
-  testEnvironment: 'node',
+  testEnvironment: 'jsdom',
   testURL: 'http://localhost',
   transform: {
     '^.+\\.(js|jsx)$': '<rootDir>/node_modules/babel-jest',