+++ server/fetchDataForRender.js
@@ -3,28 +3,25 @@ import ssrPrepass from 'react-ssr-prepass';
 import chalk from 'chalk';
 
 export const fetchDataForRender = (ServerApp, req) => {
-  let dataCache = {};
+  let data = {};
 
-  return ssrPrepass(
-    <ServerApp dataCache={dataCache} location={req.url} />,
-    element => {
-      if (element && element.type && element.type.fetchData) {
-        return element.type.fetchData(req).then(d => {
-          Object.keys(d).forEach(key => {
-            if (dataCache[key]) {
-              logDuplicateKeyMessage(key, element.type.name);
-            }
-          });
-
-          dataCache = {
-            ...dataCache,
-            ...d
-          };
+  return ssrPrepass(<ServerApp data={data} location={req.url} />, element => {
+    if (element && element.type && element.type.fetchData) {
+      return element.type.fetchData(req).then(d => {
+        Object.keys(d).forEach(key => {
+          if (data[key]) {
+            logDuplicateKeyMessage(key, element.type.name);
+          }
         });
-      }
+
+        data = {
+          ...data,
+          ...d
+        };
+      });
     }
-  ).then(() => {
-    return dataCache;
+  }).then(() => {
+    return data;
   });
 };