+++ server/renderServerSideApp.js
@@ -11,9 +11,9 @@ import { indexHtml } from './indexHtml';
 import stats from '../build/react-loadable.json';
 import { ServerDataProvider } from '../src/state/serverDataContext';
 
-const ServerApp = ({ context, dataCache, location }) => {
+const ServerApp = ({ context, data, location }) => {
   return (
-    <ServerDataProvider dataCache={dataCache}>
+    <ServerDataProvider value={data}>
       <StaticRouter location={location} context={context}>
         <App />
       </StaticRouter>
@@ -33,7 +33,7 @@ function renderApp(ServerApp, data, req, res) {
 
   const markup = ReactDOMServer.renderToString(
     <Loadable.Capture report={moduleName => modules.push(moduleName)}>
-      <ServerApp location={req.url} dataCache={data} context={context} />
+      <ServerApp location={req.url} data={data} context={context} />
     </Loadable.Capture>
   );